Improved information from `ipbb info`
