+++ src/ipbb/scripts/builder.py
@@ -13,7 +13,7 @@ from io import StringIO, BytesIO
 from texttable import Texttable
 from click import echo, style, secho
 
-from ..cmds import Environment, _utils
+from ..cmds import Environment
 from ..cmds.formatters import DepFormatter
 
 from .._version import __version__
@@ -43,81 +43,6 @@ def climain(ctx, aExcStack):
 # ------------------------------------------------------------------------------
 
 
-# ------------------------------------------------------------------------------
-@climain.command()
-@click.option('-v', '--verbose', count=True, help="Verbosity")
-@click.pass_obj
-def info(env, verbose):
-    '''Print a brief report about the current working area'''
-
-    if not env.work.path:
-        secho('ERROR: No ipbb work area detected', fg='red')
-        return
-
-    echo()
-    secho("ipbb environment", fg='blue')
-    # echo  ( "----------------")
-    lEnvTable = Texttable(max_width=0)
-    lEnvTable.add_row(["Work path", env.work.path])
-    if env.currentproj.path:
-        lEnvTable.add_row(["Project path", env.currentproj.path])
-    echo(lEnvTable.draw())
-
-    if not env.currentproj.path:
-        echo()
-        secho("Firmware packages", fg='blue')
-        lSrcTable = Texttable()
-        lSrcTable.set_deco(Texttable.HEADER | Texttable.BORDER)
-        for lSrc in env.sources:
-            lSrcTable.add_row([lSrc])
-        echo(lSrcTable.draw())
-
-        echo()
-        secho("Projects", fg='blue')
-        lProjTable = Texttable()
-        lProjTable.set_deco(Texttable.HEADER | Texttable.BORDER)
-        for lProj in env.projects:
-            lProjTable.add_row([lProj])
-        echo(lProjTable.draw())
-        return
-
-    echo()
-
-    if not env.currentproj.settings:
-        return
-
-    secho("Project '%s'" % env.currentproj.name, fg='blue')
-
-    echo(_utils.formatDictTable(env.currentproj.settings, aHeader=False))
-
-    echo()
-
-    if env.currentproj.usersettings:
-        secho("User settings", fg='blue')
-        echo(_utils.formatDictTable(env.currentproj.usersettings, aHeader=False))
-
-        echo()
-
-    lParser = env.depParser
-    lDepFmt = DepFormatter(lParser)
-    
-    if lParser.errors:
-        secho("Dep tree parsing error(s):", fg='red')
-        echo(lDepFmt.drawParsingErrors())
-        echo()
-
-    secho("Dependecy tree elements", fg='blue')
-    echo(lDepFmt.drawDeptreeCommandsSummary())
-
-    echo()
-
-    if  lParser.unresolved:
-        secho("Unresolved item(s)", fg='red')
-        echo(lDepFmt.drawUnresolvedSummary())
-
-        echo()
-# ------------------------------------------------------------------------------
-
 
 # ------------------------------------------------------------------------------
 def _compose_cli():
@@ -125,7 +50,7 @@ def _compose_cli():
     from ..cli import repo
 
     climain.add_command(repo.init)
-    # climain.add_command(repo.cd)
+    climain.add_command(repo.info)
     climain.add_command(repo.add)
     climain.add_command(repo.srcs)
 
@@ -178,54 +103,6 @@ def main():
     if sys.version_info[0:2] < (2, 7):
         click.secho("Error: Python 2.7 is required to run IPBB", fg='red')
         raise SystemExit(-1)
-
-    # if sys.version_info[0:2] == (2, 7):
-    #     click.secho(
-    #         "Warning: IPBB prefers python 3.0. python 2.7 will be deprecated in the near future.",
-    #         fg='yellow',
-    #     )
-
-    # # Add custom cli to shell
-    # from ..cli import repo
-
-    # climain.add_command(repo.init)
-    # # climain.add_command(repo.cd)
-    # climain.add_command(repo.add)
-    # climain.add_command(repo.srcs)
-
-    # from ..cli import proj
-
-    # climain.add_command(proj.proj)
-
-    # from ..cli import dep
-
-    # climain.add_command(dep.dep)
-
-    # from ..cli import toolbox
-
-    # climain.add_command(toolbox.toolbox)
-
-    # from ..cli import common
-
-    # from ..cli import vivado
-
-    # vivado.vivado.add_command(common.cleanup)
-    # vivado.vivado.add_command(common.addrtab)
-    # vivado.vivado.add_command(common.gendecoders)
-    # vivado.vivado.add_command(common.user_config)
-    # climain.add_command(vivado.vivado)
-
-    # from ..cli import sim
-
-    # sim.sim.add_command(common.cleanup)
-    # sim.sim.add_command(common.addrtab)
-    # sim.sim.add_command(common.gendecoders)
-    # sim.sim.add_command(common.user_config)
-    # climain.add_command(sim.sim)
-
-    # from ..cli import debug
-
-    # climain.add_command(debug.debug)
 
     _compose_cli()
